+++ bin/review-toc
@@ -14,6 +14,6 @@ require 'pathname'
 bindir = Pathname.new(__FILE__).realpath.dirname
 $LOAD_PATH.unshift((bindir + '../lib').realpath)
 
-require 'review/indexprinter'
+require 'review/tocprinter'
 
-ReVIEW::IndexPrinter.execute(*ARGV)
+ReVIEW::TOCPrinter.execute(*ARGV)
